cast print value to int in debug message

+++ src/kernelMidi.cpp
@@ -235,7 +235,7 @@ void callback(double t, std::vector<unsigned char> *msg, void *data) {
 	 * messages) as unknown, for debugging purposes */
 
 	if (msg->size() < 3) {
-		printf("[KM] MIDI received - unkown signal - size=%d, value=0x", msg->size());
+		printf("[KM] MIDI received - unkown signal - size=%d, value=0x", (int) msg->size());
 		for (unsigned i=0; i<msg->size(); i++)
 			printf("%X", (int) msg->at(i));
 		printf("\n");
